refactor: reorganize classes into task/, git/, and domain/ subpackages
Move classes out of the flat root package into concern-based subpackages:
- task/: PrintChangedProjectsTask, PrintChangedProjectsFromRefTask, WriteChangedProjectsFromRefTask
- git/: GitRepository, GitChangedFilesDetector, BaseBranchResolver (GitCommandExecutor already here)
- domain/: ProjectFileMapper, ProjectMetadataFactory (ProjectMetadata, ChangedProjects already here)

Update package declarations, add necessary imports, and move test files to
mirror the production package structure.
