+++ monorepo-build-plugin/src/main/kotlin/io/github/doughawley/monorepobuild/task/PrintChangedProjectsTask.kt
@@ -1,5 +1,7 @@
-package io.github.doughawley.monorepobuild
+package io.github.doughawley.monorepobuild.task
 
+import io.github.doughawley.monorepobuild.MonorepoBuildExtension
+import io.github.doughawley.monorepobuild.MonorepoBuildPlugin
 import org.gradle.api.DefaultTask
 import org.gradle.api.tasks.TaskAction
